refactor(util): merge propertyAccess and propertyName into property module
- Create lib/util/property.js with merged propertyAccess and propertyName
- Update all imports across lib/ and test/
- Remove lib/util/propertyAccess.js and lib/util/propertyName.js
