+++ lib/ExternalModule.js
@@ -25,7 +25,7 @@ const StaticExportsDependency = require("./dependencies/StaticExportsDependency"
 const createHash = require("./util/createHash");
 const extractUrlAndGlobal = require("./util/extractUrlAndGlobal");
 const makeSerializable = require("./util/makeSerializable");
-const propertyAccess = require("./util/propertyAccess");
+const { propertyAccess } = require("./util/property");
 const { register } = require("./util/serialization");
 
 /** @typedef {import("webpack-sources").Source} Source */
