+++ lib/RuntimeTemplate.js
@@ -19,7 +19,7 @@ const {
 const { equals } = require("./util/ArrayHelpers");
 const compileBooleanMatcher = require("./util/compileBooleanMatcher");
 const memoize = require("./util/memoize");
-const propertyAccess = require("./util/propertyAccess");
+const { propertyAccess } = require("./util/property");
 const { forEachRuntime, subtractRuntime } = require("./util/runtime");
 
 const getHarmonyImportDependency = memoize(() =>
