+++ lib/css/CssGenerator.js
@@ -46,7 +46,7 @@ const memoize = require("../util/memoize");
 /** @typedef {import("../Module").RuntimeRequirements} RuntimeRequirements */
 /** @typedef {import("../../declarations/WebpackOptions").CssParserExportType} CssParserExportType */
 
-const getPropertyName = memoize(() => require("../util/propertyName"));
+const getPropertyName = memoize(() => require("../util/property"));
 const getCssModulesPlugin = memoize(() => require("./CssModulesPlugin"));
 
 class CssGenerator extends Generator {
