+++ lib/dependencies/CommonJsExportRequireDependency.js
@@ -10,7 +10,7 @@ const { UsageState } = require("../ExportsInfo");
 const Template = require("../Template");
 const { equals } = require("../util/ArrayHelpers");
 const makeSerializable = require("../util/makeSerializable");
-const propertyAccess = require("../util/propertyAccess");
+const { propertyAccess } = require("../util/property");
 const { handleDependencyBase } = require("./CommonJsDependencyHelpers");
 const ModuleDependency = require("./ModuleDependency");
 const processExportInfo = require("./processExportInfo");
