+++ lib/dependencies/CommonJsExportsDependency.js
@@ -7,7 +7,7 @@
 
 const InitFragment = require("../InitFragment");
 const makeSerializable = require("../util/makeSerializable");
-const propertyAccess = require("../util/propertyAccess");
+const { propertyAccess } = require("../util/property");
 const { handleDependencyBase } = require("./CommonJsDependencyHelpers");
 const NullDependency = require("./NullDependency");
