+++ lib/dependencies/CommonJsExportsParserPlugin.js
@@ -8,7 +8,7 @@
 const RuntimeGlobals = require("../RuntimeGlobals");
 const formatLocation = require("../formatLocation");
 const { evaluateToString } = require("../javascript/JavascriptParserHelpers");
-const propertyAccess = require("../util/propertyAccess");
+const { propertyAccess } = require("../util/property");
 const CommonJsExportRequireDependency = require("./CommonJsExportRequireDependency");
 const CommonJsExportsDependency = require("./CommonJsExportsDependency");
 const CommonJsSelfReferenceDependency = require("./CommonJsSelfReferenceDependency");
