+++ lib/dependencies/CommonJsSelfReferenceDependency.js
@@ -8,7 +8,7 @@
 const RuntimeGlobals = require("../RuntimeGlobals");
 const { equals } = require("../util/ArrayHelpers");
 const makeSerializable = require("../util/makeSerializable");
-const propertyAccess = require("../util/propertyAccess");
+const { propertyAccess } = require("../util/property");
 const NullDependency = require("./NullDependency");
 
 /** @typedef {import("webpack-sources").ReplaceSource} ReplaceSource */
