+++ lib/dependencies/HarmonyExportExpressionDependency.js
@@ -8,7 +8,7 @@
 const ConcatenationScope = require("../ConcatenationScope");
 const RuntimeGlobals = require("../RuntimeGlobals");
 const makeSerializable = require("../util/makeSerializable");
-const propertyAccess = require("../util/propertyAccess");
+const { propertyAccess } = require("../util/property");
 const HarmonyExportInitFragment = require("./HarmonyExportInitFragment");
 const NullDependency = require("./NullDependency");
