+++ lib/dependencies/HarmonyExportImportedSpecifierDependency.js
@@ -18,8 +18,7 @@ const {
 const { countIterable } = require("../util/IterableHelpers");
 const { combine, first } = require("../util/SetHelpers");
 const makeSerializable = require("../util/makeSerializable");
-const propertyAccess = require("../util/propertyAccess");
-const { propertyName } = require("../util/propertyName");
+const { propertyAccess, propertyName } = require("../util/property");
 const {
 	filterRuntime,
 	getRuntimeKey,
