+++ lib/dependencies/HarmonyExportInitFragment.js
@@ -8,7 +8,7 @@
 const InitFragment = require("../InitFragment");
 const RuntimeGlobals = require("../RuntimeGlobals");
 const { first } = require("../util/SetHelpers");
-const { propertyName } = require("../util/propertyName");
+const { propertyName } = require("../util/property");
 
 /** @typedef {import("webpack-sources").Source} Source */
 /** @typedef {import("../Generator").GenerateContext} GenerateContext */
