+++ lib/dependencies/ImportMetaPlugin.js
@@ -24,7 +24,7 @@ const {
 	toConstantDependency
 } = require("../javascript/JavascriptParserHelpers");
 const memoize = require("../util/memoize");
-const propertyAccess = require("../util/propertyAccess");
+const { propertyAccess } = require("../util/property");
 const ConstDependency = require("./ConstDependency");
 
 /** @typedef {import("estree").MemberExpression} MemberExpression */
