+++ lib/library/AssignLibraryPlugin.js
@@ -9,7 +9,7 @@ const { ConcatSource } = require("webpack-sources");
 const { UsageState } = require("../ExportsInfo");
 const RuntimeGlobals = require("../RuntimeGlobals");
 const Template = require("../Template");
-const propertyAccess = require("../util/propertyAccess");
+const { propertyAccess } = require("../util/property");
 const { getEntryRuntime } = require("../util/runtime");
 const AbstractLibraryPlugin = require("./AbstractLibraryPlugin");
