+++ lib/library/ExportPropertyLibraryPlugin.js
@@ -8,7 +8,7 @@
 const { ConcatSource } = require("webpack-sources");
 const { UsageState } = require("../ExportsInfo");
 const RuntimeGlobals = require("../RuntimeGlobals");
-const propertyAccess = require("../util/propertyAccess");
+const { propertyAccess } = require("../util/property");
 const { getEntryRuntime } = require("../util/runtime");
 const AbstractLibraryPlugin = require("./AbstractLibraryPlugin");
