+++ lib/library/SystemLibraryPlugin.js
@@ -9,7 +9,7 @@ const { ConcatSource } = require("webpack-sources");
 const { UsageState } = require("../ExportsInfo");
 const ExternalModule = require("../ExternalModule");
 const Template = require("../Template");
-const propertyAccess = require("../util/propertyAccess");
+const { propertyAccess } = require("../util/property");
 const AbstractLibraryPlugin = require("./AbstractLibraryPlugin");
 
 /** @typedef {import("webpack-sources").Source} Source */
