+++ lib/optimize/ConcatenatedModule.js
@@ -45,8 +45,7 @@ const {
 const createHash = require("../util/createHash");
 const { makePathsRelative } = require("../util/identifier");
 const makeSerializable = require("../util/makeSerializable");
-const propertyAccess = require("../util/propertyAccess");
-const { propertyName } = require("../util/propertyName");
+const { propertyAccess, propertyName } = require("../util/property");
 const {
 	filterRuntime,
 	intersectRuntime,
